Implemented difference profiles for pump-probe experiments: -cumulative sums of pumped and dark images are calculated and scaled independently, and then the difference is calculated. -Individual frames are also represented as difference profiles, by subtracting scaled cumulative dark average.
Intensity sum plot is now shown as a histogram. This allows one to
more easily see the distribution of unscaled total intensities,
useful for selecting an intensity threshold filter to identify hits.

Fixed hit rate calculator to be correct even when the running window
was larger than the number of events collected thus far, e.g. at the
beginning of the experiment.
- X-axis does not properly scale to bins. Need to fix this. Attempts
to do this failed, resulting in rescaling the axis to improper values.
- currently just defaults to the number of bins (e.g. 100), from which
the intensities can be estimated based on the minimum and maximum
values given by the user in the monitor.ini file.